Organize contexts into subfolders
